renamed entities and new make recipe
