+++ api/src/modules/planning-units/dto/create.planning-units.dto.ts
@@ -1,1 +1,6 @@
-export class CreatePlanningUnitsDTO {}
+import { PickType } from '@nestjs/swagger';
+import { CreateProjectDTO } from 'modules/projects/dto/create.project.dto';
+
+export class CreatePlanningUnitsDTO extends PickType(CreateProjectDTO,['extent','countryId','adminAreaLevel1Id','adminAreaLevel2Id','planningUnitAreakm2','planningUnitGridShape'] as const) {
+
+}
