+++ geoprocessing/src/modules/planning-units-area-protected/planning-units-area-protected.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { CalculatePlanningUnitAreaProtectedProcessor } from './planning-units.area-protected.worker';
+import { CalculatePlanningUnitAreaProtectedProcessor } from './planning-units-area-protected.worker';
 
 /**
  * @todo: this needs to be imported at application level
